src/server/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
beanie==1.25.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
httptools==0.6.1
idna==3.6
lazy-model==0.2.0
motor==3.3.2
passlib==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.35.1
toml==0.10.2
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.27.0
watchfiles==0.21.0
websockets==12.0
